# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "An API used to manage payments with friends for home poker games :)",
        "title": "Poker Finance API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/users": {
            "post": {
                "description": "Create a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create a new user",
                "parameters": [
                    {
                        "description": "User Input",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.CreateUserInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "Get a user by ID",
                "tags": [
                    "users"
                ],
                "summary": "Get a user by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.User"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update user information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update user information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User Update Input",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.UpdateUserInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}/played-games": {
            "get": {
                "description": "Get played games by user ID",
                "tags": [
                    "users"
                ],
                "summary": "Get played games by user ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/backend_internal_models.PlayedGames"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}/played-games/{gameid}": {
            "put": {
                "description": "Update a played game's details by user ID and game ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update a played game by user ID and game ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Game ID",
                        "name": "gameid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Played Game Update Input",
                        "name": "playedGame",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.UpdateUserPlayedGameInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.PlayedGames"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new played game entry for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create a played game for a user",
                "parameters": [
                    {
                        "description": "Played Game Input",
                        "name": "playedGame",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.CreateUserPlayedGameInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.PlayedGames"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}/transactions": {
            "get": {
                "description": "Get all transactions by user ID",
                "tags": [
                    "users"
                ],
                "summary": "Get transactions by user ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/backend_internal_models.Transaction"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "backend_internal_models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "backend_internal_models.Game": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "backend_internal_models.PaymentDetails": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "details": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "payee": {
                    "$ref": "#/definitions/backend_internal_models.User"
                },
                "payeeID": {
                    "type": "string"
                },
                "payer": {
                    "$ref": "#/definitions/backend_internal_models.User"
                },
                "payerID": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "timeCompleted": {
                    "type": "string"
                },
                "timeSubmitted": {
                    "type": "string"
                }
            }
        },
        "backend_internal_models.PlayedGames": {
            "type": "object",
            "properties": {
                "buyIn": {
                    "type": "integer"
                },
                "endAmount": {
                    "type": "integer"
                },
                "game": {
                    "$ref": "#/definitions/backend_internal_models.Game"
                },
                "gameID": {
                    "type": "string"
                },
                "player": {
                    "$ref": "#/definitions/backend_internal_models.User"
                },
                "playerID": {
                    "type": "string"
                }
            }
        },
        "backend_internal_models.Transaction": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "paymentDetails": {
                    "$ref": "#/definitions/backend_internal_models.PaymentDetails"
                },
                "paymentDetailsID": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/backend_internal_models.User"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "backend_internal_models.User": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "internal_api_handlers.CreateUserInput": {
            "type": "object",
            "required": [
                "balance",
                "email",
                "name"
            ],
            "properties": {
                "balance": {
                    "type": "integer"
                },
                "display_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "internal_api_handlers.CreateUserPlayedGameInput": {
            "type": "object",
            "required": [
                "buy_in",
                "game_id",
                "user_id"
            ],
            "properties": {
                "buy_in": {
                    "type": "integer"
                },
                "game_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "internal_api_handlers.UpdateUserInput": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer"
                },
                "display_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "internal_api_handlers.UpdateUserPlayedGameInput": {
            "type": "object",
            "properties": {
                "buy_in": {
                    "type": "integer"
                },
                "end_amount": {
                    "type": "integer"
                }
            }
        }
    }
}